+++ CHANGELOG.md
@@ -113,6 +113,7 @@ Compatibility:
 * Define all C-API `rb_{c,m,e}*` constants as C global variables (#1541).
 * Raise `ArgumentError` for `Socket.unpack_sockaddr_un` if the socket family is incorrect.
 * Implemented `RTYPEDDATA_*()` macros and `rb_str_tmp_new()` (#1975).
+* Implemented `rb_set_end_proc` (#1959).
 
 Changes: